Dockerfiles/purge-druid-datasources/requirements.txt
backoff
dataclasses-json
psycopg2-binary
requests